+++ ocr/pymupdf/pymupdf_plain/ASUCapstoneDesign.txt
