+++ ocr/pymupdf/pymupdf_plain/CE438W.txt
