+++ ocr/pymupdf/pymupdf_plain/MECH-45X-Syllabus-2019.txt
