+++ ocr/tesseract/tesseract_plain/ASUCapstoneDesign.txt
