+++ ocr/tesseract/tesseract_plain/CS4704Proposal.txt
