+++ ocr/tesseract/tesseract_plain/CapstoneSyllabus2017.txt
