+++ ocr/tesseract/tesseract_plain/Capstone_Syllabus_Summer_2.txt
