+++ ocr/tesseract/tesseract_plain/MECH-45X-Syllabus-2019.txt
