+++ ocr/tesseract/tesseract_plain/TAMUT_Capstone.txt
