+++ ocr/tesseract/tesseract_plain/syllabus-ce 495B.txt
