+++ ocr/tesseract/tesseract_plain/tesseract_ocr.py
