+++ ocr/tesseract/tesseract_plain/tesseract_scores.xlsx
