+++ ocr/tesseract/tesseract_table/1-syllabus.txt
