+++ ocr/tesseract/tesseract_table/APS111.txt
