+++ ocr/tesseract/tesseract_table/APS112.txt
