+++ ocr/tesseract/tesseract_table/CIVE-ENVE-4918.txt
