--- pymupdf_plain/summary.txt
@@ -1,11 +1,0 @@
-Capstone_Syllabus_Summer_2.pdf: CER = 4.48%, WER = 0.00%
-MECH-45X-Syllabus-2019.pdf: CER = 3.10%, WER = 0.07%
-CS4704Proposal.pdf: CER = 6.37%, WER = 0.35%
-ASUCapstoneDesign.pdf: CER = 2.60%, WER = 7.29%
-CapstoneSyllabus2017.pdf: CER = 3.04%, WER = 0.00%
-TAMUT_Capstone.pdf: CER = 7.21%, WER = 0.21%
-ELT-495_syllabus_ELT-495.pdf: CER = 0.96%, WER = 11.92%
-CE438W.pdf: CER = 5.97%, WER = 0.00%
-syllabus-ce 495B.pdf: CER = 11.26%, WER = 0.00%
-Overall CER = 4.07%
-Overall WER = 3.27%
